refactor(adapter): separate concerns into distinct modules
- Move UserModel to models/user.rs (Adaptee)
- Extract services into services/user_service.rs (Target/Adapter)
- Move controller to controllers/user_controller.rs (Client)
- Add module structure for better maintainability
