+++ design-pattern/adapter/Cargo.toml
@@ -5,5 +5,3 @@ edition = "2024"
 
 [dependencies]
 chrono = { version = "0.4", features = ["serde"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
