+++ design-pattern/adapter/src/controllers/mod.rs
@@ -1,0 +1,2 @@
+pub mod user_controller;
+pub use user_controller::UserController;
