+++ design-pattern/adapter/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod models;
+pub mod services;
+pub mod controllers;
