+++ design-pattern/adapter/src/models/mod.rs
@@ -1,0 +1,2 @@
+pub mod user;
+pub use user::UserModel;
