+++ design-pattern/adapter/src/services/mod.rs
@@ -1,0 +1,2 @@
+pub mod user_service;
+pub use user_service::{UserService, UserServiceAdapter};
